[gitconfig] Add `roh` alias for `rebase origin/${HEAD}`
This alias is supposed to replace the manual calling of `git rebase
origin/master` which is how I typically update my local checkout branch
against a remote, the idea being that a rebase against an upstream
master should be a straight forward fast forward. Otherwise, and
un-pushed commits would be automatically rebased against the fast
forward.

To handle the case where we're updating a non-master branch, we'd like
the equivalent of a `git rebase origin/HEAD`, except we want the branch
name of `HEAD` not the `HEAD` literal itself. This alias resolves the
current value of `HEAD` to get something like `git rebase
origin/my-branch`.

There are 2 optional arguments to the alias, the first being an override
of the `origin` remote and the second being an override of the remote
branch name. However, overriding both of these would be super unusual
and calling `git rebase` directory is likely a much safer action.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>