[gitconfig] Update `poh` alias to take the remote & branch
This change lets a user optionally override the remote when pushing
changes. By default, the remote is `origin` and the branch is the value
`HEAD`. For example, if changes need to be pushed to a remote called
`upstream` (say, if you cloned a personal fork but still have commit
access to the upstream/canonical repository), you can use `git poh
upstream` which effectively runs `git push upstream HEAD`.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>