[tmux] start using a 1-based index for windows & panes
It is time. A zero-based indexing for windows and panes has worked for
me so far, especially with the additional tweak of adding the backtick
key as an alternative for zero as this key is *typically* to the left of
the one key. However, not for all keyboards and keyboard layouts (hint:
I may or may not be working with a new Kinesis Advantage2 keyboard), so
a more consistent experience will be to start with the first window and
pane being at 1 instead of 0.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>