feat(git): add "upstream" remote aliases
This change adds 3 additional Git aliases:

- `puh`: push upstream HEAD
- `ruh`: rebase upstream HEAD
- `rum`: rebase upstream main

These are functional identical to the `poh`/`roh`/`rom` aliases
respectively, except they operate on an "upstream" remote rather than
the default "origin" remote. Note that this assumes an "upstream" remote
has been set up for the repository, however this is a common pattern
used by the author when tracking an upstream project (i.e. working on a
downstream fork).

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>